rename add_partition.py to partition.py
