+++ subiquity/ui/views/filesystem/__init__.py
@@ -20,7 +20,7 @@ configuration.
 
 """
 
-from .add_partition import FormatEntireView, PartitionView
 from .disk_info import DiskInfoView
 from .disk_partition import DiskPartitionView
 from .filesystem import FilesystemView
+from .partition import FormatEntireView, PartitionView
