+++ subiquity/ui/views/filesystem/partition.py
